+++ src/norm/postgres.nim
@@ -10,12 +10,12 @@ PostgreSQL Backend
 import strutils, macros, typetraits, logging
 import db_postgres
 
-import rowutils, objutils, pragmas
+import oldrowutils, objutils, pragmas
 
 
 export strutils, macros, logging
 export db_postgres
-export rowutils, objutils, pragmas
+export oldrowutils, objutils, pragmas
 
 proc `$`*(query: SqlQuery): string = $ string query
 
@@ -30,7 +30,7 @@ proc getTable*(objRepr: ObjRepr): string =
     if prag.name == "table" and prag.kind == pkKval:
       return $prag.value
 
-proc getTable*(T: type): string =
+proc getTable*(T: typedesc): string =
   ##[ Get the name of the DB table for the given type: ``table`` pragma value if it exists
   or lowercased type name otherwise.
   ]##
@@ -178,16 +178,17 @@ template genWithDb(connection, user, password, database: string,
                     tableSchemas, dropTableQueries: openArray[SqlQuery]): untyped {.dirty.} =
   ## Generate ``withDb`` template.
 
-  template withDb*(body: untyped): untyped {.dirty.} =
-    ##[ A wrapper for actions that require DB connection. Defines CRUD procs to work with the DB,
-    as well as ``createTables`` and ``dropTables`` procs.
+  template withCustomDb*(customConnection, customUser, customPassword, customDatabase: string,
+                         body: untyped): untyped {.dirty.} =
+    ##[ A wrapper for actions that require custom DB connection, i.e. not the one defined in ``db``.
+    Defines CRUD procs to work with the DB, as well as ``createTables`` and ``dropTables`` procs.
 
-      Aforementioned procs and procs defined in a ``db`` block can be used only
-      in  a ``withDb`` block.
+    Aforementioned procs and procs defined in a ``db`` block can be used only
+    in  a ``withDb`` block.
     ]##
 
     block:
-      let dbConn = open(connection, user, password, database)
+      let dbConn = open(customConnection, customUser, customPassword, customDatabase)
 
       template dropTables() {.used.} =
         ## Drop tables for all types in all type sections under ``db`` macro.
@@ -244,7 +245,7 @@ template genWithDb(connection, user, password, database: string,
 
         row.to(obj)
 
-      proc getOne(T: type, cond: string, params: varargs[string, `$`]): T {.used.} =
+      proc getOne(T: typedesc, cond: string, params: varargs[string, `$`]): T {.used.} =
         ##[ Read a record from DB by condition into a new object instance.
 
         If multiple records are found, return the first one.
@@ -266,7 +267,7 @@ template genWithDb(connection, user, password, database: string,
 
         row.to(obj)
 
-      proc getOne(T: type, id: int): T {.used.} =
+      proc getOne(T: typedesc, id: int): T {.used.} =
         ## Read a record from DB into a new object instance.
 
         result.getOne(id)
@@ -290,7 +291,7 @@ template genWithDb(connection, user, password, database: string,
 
         rows.to(objs)
 
-      proc getMany(T: type, limit: int, offset = 0,
+      proc getMany(T: typedesc, limit: int, offset = 0,
                    cond = "TRUE", params: varargs[string, `$`]): seq[T] {.used.} =
         ##[ Read ``limit`` records  with ``offset`` from DB into a sequence of objects,
         create the sequence on the fly.
@@ -328,6 +329,17 @@ template genWithDb(connection, user, password, database: string,
 
       try: body
       finally: dbConn.close()
+
+  template withDb*(body: untyped): untyped {.dirty.} =
+    ##[ A wrapper for actions that require DB connection. Defines CRUD procs to work with the DB,
+    as well as ``createTables`` and ``dropTables`` procs.
+
+      Aforementioned procs and procs defined in a ``db`` block can be used only
+      in  a ``withDb`` block.
+    ]##
+
+    withCustomDb(connection, user, password, database):
+      body
 
 proc ensureIdFields(typeSection: NimNode): NimNode =
   ## Check if ``id`` field is in the object definition, insert it if it's not.