+++ tests/testsqlite.nim
@@ -5,14 +5,21 @@ import os, strutils, sequtils, times
 import norm / sqlite
 
 
-db("test.db", "", "", ""):
+const
+  dbName = "test.db"
+  customDbName = "custom_test.db"
+
+
+db(dbName, "", "", ""):
   type
     User {.table: "users".} = object
       email {.unique.}: string
+      ssn: Option[int]
       birthDate {.
         dbType: "INTEGER",
-        parseIt: it.parseInt().fromUnix().local(),
-        formatIt: $it.toTime().toUnix()
+        notNull,
+        parseIt: it.i.fromUnix().local(),
+        formatIt: dbValue(it.toTime().toUnix())
       .}: DateTime
     Publisher {.table: "publishers".} = object
       title {.unique.}: string
@@ -21,7 +28,7 @@ db("test.db", "", "", ""):
       authorEmail {.fk: User.email, onDelete: "CASCADE".}: string
       publisherTitle {.fk: Publisher.title.}: string
 
-  proc getBookById(id: string): Book = withDb(Book.getOne parseInt(id))
+  proc getBookById(id: DbValue): Book = withDb(Book.getOne int(id.i))
 
   type
     Edition {.table: "editions".} = object
@@ -29,9 +36,10 @@ db("test.db", "", "", ""):
       book {.
         dbCol: "bookId",
         dbType: "INTEGER",
+        notNull,
         fk: Book
         parser: getBookById,
-        formatIt: $it.id,
+        formatIt: dbValue(it.id),
         onDelete: "CASCADE"
       .}: Book
 
@@ -42,7 +50,7 @@ suite "Creating and dropping tables, CRUD":
 
       for i in 1..9:
         var
-          user = User(email: "test-$#@example.com" % $i,
+          user = User(email: "test-$#@example.com" % $i, ssn: some i,
                       birthDate: parse("200$1-0$1-0$1" % $i, "yyyy-MM-dd"))
           publisher = Publisher(title: "Publisher $#" % $i)
           book = Book(title: "Book $#" % $i, authorEmail: user.email,
@@ -62,23 +70,24 @@ suite "Creating and dropping tables, CRUD":
 
   test "Create tables":
     withDb:
-      let query = sql "PRAGMA table_info(?);"
+      let query = "PRAGMA table_info($#);"
 
-      check dbConn.getAllRows(query, "users") == @[
-        @["0", "id", "INTEGER", "0", "", "1"],
-        @["1", "email", "TEXT", "0", "", "0"],
-        @["2", "birthDate", "INTEGER", "0", "", "0"]
+      check dbConn.getAllRows(sql query % "users") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "ssn", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "birthDate", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
       ]
-      check dbConn.getAllRows(query, "books") == @[
-        @["0", "id", "INTEGER", "0", "", "1"],
-        @["1", "title", "TEXT", "0", "", "0"],
-        @["2", "authorEmail", "TEXT", "0", "", "0"],
-        @["3", "publisherTitle", "TEXT", "0", "", "0"],
+      check dbConn.getAllRows(sql query % "books") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
       ]
-      check dbConn.getAllRows(query, "editions") == @[
-        @["0", "id", "INTEGER", "0", "", "1"],
-        @["1", "title", "TEXT", "0", "", "0"],
-        @["2", "bookId", "INTEGER", "0", "", "0"]
+      check dbConn.getAllRows(sql query % "editions") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
       ]
 
   test "Create records":
@@ -161,7 +170,7 @@ suite "Creating and dropping tables, CRUD":
   test "Query records":
     withDb:
       let someBooks = Book.getMany(10, cond="title IN (?, ?) ORDER BY title DESC",
-                                   params=["Book 1", "Book 5"])
+                                   params=[dbValue "Book 1", dbValue "Book 5"])
 
       check len(someBooks) == 2
       check someBooks[0].title == "Book 5"
@@ -214,4 +223,40 @@ suite "Creating and dropping tables, CRUD":
         dbConn.exec sql "SELECT NULL FROM books"
         dbConn.exec sql "SELECT NULL FROM editions"
 
-  removeFile "test.db"
+  test "Custom DB":
+    withCustomDb(customDbName, "", "", ""):
+      createTables(force=true)
+
+    withCustomDb(customDbName, "", "", ""):
+      let query = "PRAGMA table_info($#);"
+
+      check dbConn.getAllRows(sql query % "users") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "email", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "ssn", dbValue "INTEGER", dbValue 0, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "birthDate", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
+      ]
+      check dbConn.getAllRows(sql query % "books") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "authorEmail", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 3, dbValue "publisherTitle", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+      ]
+      check dbConn.getAllRows(sql query % "editions") == @[
+        @[dbValue 0, dbValue "id", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 1],
+        @[dbValue 1, dbValue "title", dbValue "TEXT", dbValue 1, dbValue nil, dbValue 0],
+        @[dbValue 2, dbValue "bookId", dbValue "INTEGER", dbValue 1, dbValue nil, dbValue 0]
+      ]
+
+    withCustomDb(customDbName, "", "", ""):
+      dropTables()
+
+      expect DbError:
+        dbConn.exec sql "SELECT NULL FROM users"
+        dbConn.exec sql "SELECT NULL FROM publishers"
+        dbConn.exec sql "SELECT NULL FROM books"
+        dbConn.exec sql "SELECT NULL FROM editions"
+
+    removeFile customDbName
+
+  removeFile dbName